Revise predicate caching tests
