+++ src/main/java/org/javarosa/xpath/expr/XPathEqExpr.java
@@ -16,18 +16,17 @@
 
 package org.javarosa.xpath.expr;
 
-import java.io.DataInputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
-
 import org.javarosa.core.model.condition.EvaluationContext;
 import org.javarosa.core.model.instance.DataInstance;
 import org.javarosa.core.util.externalizable.DeserializationException;
 import org.javarosa.core.util.externalizable.ExtUtil;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
-import org.javarosa.measure.Measure;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
 
 import static java.lang.Math.abs;
 import static org.javarosa.xpath.expr.XPathFuncExpr.toBoolean;
@@ -51,8 +50,6 @@ public class XPathEqExpr extends XPathBinaryOpExpr {
 
     @Override
     public Object eval(DataInstance model, EvaluationContext evalContext) {
-        Measure.log("XPathEqExpr#eval");
-
         logger.debug("XPathEqExpr{}.eval starting. model: {}, candidate: {}, expecting equal: {}", id(), model,
             evalContext.candidateValue == null ? "None" : evalContext.candidateValue.getDisplayText(), equal);
         final Object aval = unpack(a.eval(model, evalContext));
